Update dialog buffix

+++ jobseeker-backend/src/controllers/application.controller.ts
@@ -142,7 +142,7 @@ export class ApplicationController {
   async updateById(
     @param.path.string('id') id: string,
     @requestBody()
-    jobApplication: JobApplication,
+    jobApplication: any,
   ): Promise<any> {
     try {
       try {
@@ -151,8 +151,6 @@ export class ApplicationController {
         return this.responseObject.customResponse(true, "Invalid Session", 401);
       }
       const fields = {
-        'id': 'string',
-        'description': 'string',
         'user': 'number',
         'company': 'string',
         'position': 'string',
@@ -182,8 +180,8 @@ export class ApplicationController {
       return this.responseObject.customResponse(true, "Invalid Session", 401);
     }
     try {
-      const application = await this.dataServiceProvider.checkUserAccessToApplication(this.request, id);
-      this.jobApplicationRepository.deleteById(id);
+      await this.dataServiceProvider.checkUserAccessToApplication(this.request, id);
+      await this.jobApplicationRepository.deleteById(id);
       return this.responseObject.successResponse();
     } catch (error) {
       return this.responseObject.customResponse(true, "There was an error while processing the request", 500);
@@ -202,7 +200,6 @@ export class ApplicationController {
       return this.responseObject.customResponse(true, "Invalid Session", 401);
     }
     try {
-
       user = await this.sessionServiceProvider.getUserFromToken(this.request.headers.authentication);
       const userString = `${user.name[0].toUpperCase()}.${user.last_name}`;
       const workbook = new Excel.Workbook();
